Script Compiler: Improve Debugging
The compilation process is now cancelled when parsing or lexing errors have occured.
The editor now places the caret near the offending text and scrolls to the location.
Added additional information to the logger.
Tried to make the compiler exception messages a bit clearer.
Special functions throw exceptions if they have an unexpected number of arguments.
Reduced the amount of special value type strings. They are now stored in a static helper class. The only remaining occurence is in a switch statement.