Remove external expression blend references
Drawing was only used for the triangles on each page, which has been replaced with the standard polygon control. Also took this opportunity to try and standardize the left side of pages more. (HaloMap's header is now matches every other page)
Interactions doesnt actually seem to be used anywhere so its also gone.