Make save all open tabs feature actually save all
this weird ass implementation for the fallback plugin path didnt need to exist i dont think. if the tag didnt have a plugin in the initial path (so Halo3MCC for example, where anything missing falls back to Halo3), it would be skipped as i was checking for only that path.