Add ability to dump an open tag as a text file
No intention to even attempt reading from these text files, fyi. This is to make comparisons between builds or mods easier. (plugins permitting as they are used to generate the output)
This also simplifies the bottom tag editor row; options is now a gear and the revisions button (along with this dump command) is within that popup menu.