Slight cleanup of engines.xml
There are no longer 2 versions of the H1 MCC scripting xml, as that isn't the case for other engines. A new "full" engine entry has been added for each MCC title at particular points to reduce redundancy in later inherits. Stub string id xmls have been added to go along with the MCC updates that put namespace information in the map header. Added a missing attribute to the guide at the top of the engines xml, as well as noting boolean defaults.