Implement H1 Plugin Research From Invader
Generated similar entity plugins from Guerilla to diff against Invader's provided 52.1 set as a guide. This gives names to the build-time written fields that was previously listed as Unknown due to not being defined in Guerilla. While H1 was the focus H2 and later games benefit as well. Thanks Aerocatia and Snowy Mouse.