Add source enum to ITag
So it can be known at a glance where a tag is without trying to search around. Used now so BSP tags can write 0s back to the tag table, but could be checked by the UI to dynamically display a "warning" when something isnt in the main metaarea. Also cleans up firstgentag in regards to custom edition resources.